requirements.txt
python-dotenv
rich
peewee
spotipy
psycopg2